feat: schedule 도메인 예외 처리

+++ src/main/java/com/project/coalba/domain/schedule/service/BossScheduleService.java
@@ -2,13 +2,13 @@ package com.project.coalba.domain.schedule.service;
 
 import com.project.coalba.domain.profile.entity.Staff;
 import com.project.coalba.domain.profile.service.StaffProfileService;
-import com.project.coalba.domain.schedule.entity.enums.TotalScheduleStatus;
-import com.project.coalba.domain.schedule.entity.enums.ScheduleStatus;
+import com.project.coalba.domain.schedule.entity.enums.*;
 import com.project.coalba.domain.schedule.service.dto.*;
 import com.project.coalba.domain.schedule.entity.Schedule;
 import com.project.coalba.domain.schedule.repository.ScheduleRepository;
 import com.project.coalba.domain.workspace.entity.Workspace;
 import com.project.coalba.domain.workspace.service.BossWorkspaceService;
+import com.project.coalba.global.exception.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -97,6 +97,7 @@ public class BossScheduleService {
         return staffProfileService.getStaffListInWorkspaceAndPossibleForDateTimeRange(workspaceId, fromDateTime, toDateTime);
     }
 
+    //TODO: INVALID_SCHEDULE_WORKER 예외 처리
     @Transactional
     public void save(ScheduleCreateServiceDto serviceDto) {
         Workspace workspace = bossWorkspaceService.getWorkspace(serviceDto.getWorkspaceId());
@@ -107,6 +108,9 @@ public class BossScheduleService {
 
     @Transactional
     public void cancel(Long scheduleId) {
+        Schedule schedule = scheduleRepository.findById(scheduleId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.SCHEDULE_NOT_FOUND));
+        if (schedule.getStatus() != ScheduleStatus.BEFORE_WORK) throw new BusinessException(ErrorCode.INVALID_SCHEDULE_CANCEL);
         scheduleRepository.deleteById(scheduleId);
     }
 }
